% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_adj_rmses.R
\name{compute_adj_rmses}
\alias{compute_adj_rmses}
\title{Function for getting "adjusted" RMSEs}
\usage{
compute_adj_rmses(
  index_adj,
  var_group,
  rep,
  type,
  y_obs,
  model,
  blockSize,
  y_spatial,
  phi,
  weights,
  scale_y,
  return_adj_preds
)
}
\arguments{
\item{index_adj}{Integer specifying which time should be permuted or
set to 0}

\item{var_group}{Vectors of integers indicating the numbers of columns
that should be included in the permuting/zeroing process}

\item{rep}{Used only for keeping track of reps associated with 
`compute_fi`.}

\item{type}{Feature importance type to be computed ('pfi' or 'zfi')}

\item{y_obs}{Matrix of y values as computed internally in compute_zfi}

\item{model}{Model output from fit_esn function}

\item{blockSize}{number of time points to permute or zero out}

\item{y_spatial}{Matrix of observed response values on the spatial scale with
rows corresponding to times and columns corresponding to spatial 
locations (should be the same matrix used for Ztrain in compute_eofs).}

\item{phi}{Matrix object phi output from compute_eofs when applied to Ztrain.}

\item{weights}{Vector of weights with length corresponding to the number of 
columns in y (or y_spatial, if y_spatial is specified) Note: currently
assumes that the weights are equivalent across time -- only varies by
location.}

\item{scale_y}{Indicates whether y and corresponding predictions should be
scaled before computing RMSEs. Scaling y values and predictions is 
intended to put multiple response variables on the same scale 
for comparison in the computation of PFI. (Default is FALSE.)}

\item{return_adj_preds}{Indicates whether the permuted/zeroed predictions are
returned in addition to PFI/ZFI values. Default is FALSE.}
}
\description{
Sets the variables values for the specified time to 0 and obtains ESN
predictions with the "adjusted" times and computes RMSEs
}
