% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eofs.R
\name{compute_eofs}
\alias{compute_eofs}
\title{Compute EOFs}
\usage{
compute_eofs(Ztrain, Ztest = NULL, n_eofs)
}
\arguments{
\item{Ztrain}{Matrix of spatio-temporal training data with T (no. times) 
rows and S (no. spatial locations) columns (rows must be
ordered top to bottom from earliest to latest observation)}

\item{Ztest}{(Optional) Matrix of spatio-temporal testing data with T 
(no. times) rows and S (no. spatial locations) columns (rows 
must be ordered top to bottom from earliest to latest 
observation) Set to NULL by default.}

\item{n_eofs}{Number of EOFs to return}
}
\description{
Function for computing EOFs given training and (optional) testing data. 
Test data principal component time series are computed using the 
training data EOFs.
}
\examples{
# Create training data
x = matrix(c(rnorm(12,10,1), rnorm(12,0,1)), ncol = 2, byrow = FALSE)
y = matrix(c(rnorm(12,5,3), rnorm(12,5,3)), ncol = 2, byrow = FALSE)

# Create a vector of times
t = c("Jan", "Feb", "Mar", "Apr", "May", "Jun",
      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
t = paste0(t, "2021")

# Create test data
x_new = matrix(c(rnorm(5,10,1), rnorm(5,0,1)), ncol = 2, byrow = FALSE)
t_new = c("Jan", "Feb", "Mar", "Apr", "May")
t_new = paste0(t_new, "2022")

# Compute EOFs
eofs <-
  compute_eofs(
    Ztrain = x,
    Ztest = x_new,
    n_eofs = 2
   )
}
