% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fi.R
\name{compute_fi}
\alias{compute_fi}
\title{Function for computing FI with ESN}
\usage{
compute_fi(
  model,
  type,
  nreps = NULL,
  var_groups = NULL,
  blockSize = 1,
  y_spatial = NULL,
  phi = NULL,
  weights = NULL,
  scale_y = FALSE,
  return_adj_preds = FALSE,
  seed = NULL
)
}
\arguments{
\item{model}{Object output from fit_esn or fit_Eesn function}

\item{type}{Feature importance type to be computed ('pfi' or 'zfi')}

\item{nreps}{Number of reps to use when computing PFI.}

\item{var_groups}{List of vectors where each vectors contains columns numbers
indicating which x variables should be grouped when doing the
zeroing process (set to NULL by default which zeros all x variables)}

\item{blockSize}{Total number of time points to zero out, defaults to 1}

\item{y_spatial}{Matrix of observed response values on the spatial scale with
rows corresponding to times and columns corresponding to spatial 
locations (should be the same matrix used for Ztrain in compute_eofs).}

\item{phi}{Matrix object phi output from compute_eofs when applied to Ztrain.}

\item{weights}{Vector of weights with length corresponding to the number of 
columns in y (or y_spatial, if y_spatial is specified) Note: currently
assumes that the weights are equivalent across time -- only varies by
location.}

\item{scale_y}{Indicates whether y and corresponding predictions should be
scaled before computing RMSEs. Scaling y values and predictions is 
intended to put multiple response variables on the same scale 
for comparison in the computation of FI. (Default is FALSE.)}

\item{return_adj_preds}{Indicates whether the permuted/zeroed predictions are
returned in addition to PFI/ZFI values. Default is FALSE.}

\item{seed}{Random seed (NULL by default.)}
}
\description{
Computes FI using the training data specified in fit_esn
}
\examples{
# Create data
x = matrix(rnorm(40,0,1), ncol = 4)
y = matrix(rnorm(20,0,1), ncol = 2)

# Assign column names
colnames(x) = c("x11", "x12", "x21", "x22")
colnames(y) = c("y1", "y2")

# Times
t = paste0("t", 1:10)

# Fit an ESN model
esn <- 
  fit_esn(
    x = x, 
    y = y, 
    t = t, 
    tau = 2, 
    m = 1, 
    tau_emb = 1, 
    nh = 10, 
    seed = 1020349858
  )

# Compute ZFI
zfi <- 
  compute_fi(
    model = esn, 
    type = "zfi",
    var_groups = list(1:2, 3:4), 
    seed = 10203498,
    blockSize = 1
   )
   
# Compute ZFI
pfi <- 
  compute_fi(
    model = esn, 
    type = "pfi",
    nreps = 2,
    var_groups = list(1:2, 3:4), 
    seed = 10203498,
    blockSize = 1
   )
}
