% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_objs.R
\name{create_data_obj}
\alias{create_data_obj}
\alias{create_data_obj_x}
\alias{create_data_obj_y}
\alias{create_data_obj_oos}
\alias{create_data_obj_ood}
\alias{create_data_obj_new}
\title{Create embedding and response matrices}
\usage{
create_data_obj(x, y, t, tau, m, tau_emb, internal_scaling)

create_data_obj_x(x, t, tau, m, tau_emb, internal_scaling)

create_data_obj_y(y, t, tau, m, tau_emb, internal_scaling)

create_data_obj_oos(model)

create_data_obj_ood(model, x_ood, t_ood)

create_data_obj_new(model, x_new, t_new)
}
\arguments{
\item{x}{Training data inputs}

\item{y}{Training data outputs}

\item{t}{Vector containing times/dates associated with x and y matrices}

\item{tau}{Forecast lead time}

\item{m}{Embedding vector length}

\item{tau_emb}{Embedding lag}

\item{internal_scaling}{Specifies the type of standardization that should be 
applied internally to x and y. This is common in ESN literature.
Options are "joint" (the mean and standard deviation from all values 
in the x/y matrix are computed after appropriate times are removed
and used to center and scale the values in the x/y matrix) and "none"
(no internal scaling is applied). Default is "joint".}

\item{model}{ESN model output from fit_esn}

\item{x_ood}{Out of distribution data inputs (must be the same size matrix
as training data x and occur at the same times as x)}

\item{t_ood}{Vector of times associated with x_ood matrix}

\item{x_new}{Data input matrix occurring after the times in the training
data x}

\item{t_new}{Vector of times associated with x_new matrix}
}
\description{
This function creates a data object containing the matrix of embedding
vectors and the corresponding response variable matrix. Scaling is done
because Wikle et al. (2019) say "the responses and inputs are scaled by 
their respective standard deviations, as is common in the ESN literature".
}
\examples{
# Create data
x = matrix(c(rnorm(12,10,1), rnorm(12,0,1)), ncol = 2, byrow = FALSE)
y = matrix(x[,1], ncol = 1)

# Assign column names
colnames(x) = c("X1", "X2")
colnames(y) = "X1"

# Create a vector of times
t = c("Jan", "Feb", "Mar", "Apr", "May", "Jun", 
      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
t = paste0(t, "2021")

# Tuning parameters
tau = 2
m = 1
tau_emb = 1
internal_scaling = TRUE

# Create data obj
data_obj <- 
  create_data_obj(
    x = x, 
    y = y, 
    t = t, 
    tau = tau, 
    m = m, 
    tau_emb = tau_emb, 
    internal_scaling = internal_scaling
 )
 
}
