% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_Eesn.R
\name{fit_Eesn}
\alias{fit_Eesn}
\title{Fit Ensemble ESN model}
\usage{
fit_Eesn(
  x,
  y,
  t,
  tau,
  m,
  tau_emb,
  nh,
  h_start = NULL,
  U_width = 0.1,
  W_width = 0.1,
  U_pi = 0.1,
  W_pi = 0.1,
  nu = 0.35,
  reg_par = 0.01,
  add_quad = TRUE,
  internal_scaling = "joint",
  seed,
  n_ensembles,
  cores = 1
)
}
\arguments{
\item{x}{Matrix of training data inputs with T rows (number of times) and 
P columns (number of input variables) - Note: rows must be ordered 
top to bottom from earliest to latest observation)}

\item{y}{Matrix of training data outputs with T rows (number of times) and 
Q columns (number of response variables) - Note: times associated
with the rows must match those in x}

\item{t}{Vector containing times/dates associated with x and y matrices}

\item{tau}{Forecast lead time}

\item{m}{Embedding vector length}

\item{tau_emb}{Embedding lag}

\item{nh}{Number of hidden units}

\item{h_start}{Values to use for initializing h (vector of length nh)
(If NULL, will be set to a vector of 0s)}

\item{U_width}{Uniform distribution sampling range for U ("width" parameter)
(Default is .1)}

\item{W_width}{Uniform distribution sampling range for W ("width" parameter)
(Default is .1)}

\item{U_pi}{Probabilities of non-zeros in U (sparseness parameter)
(Default is .1)}

\item{W_pi}{Probabilities of non-zeros in W (sparseness parameter)
(Default is .1)}

\item{nu}{Scaling parameter for W-weight matrix) (Default is .35)}

\item{reg_par}{Output ridge regression parameter (Default is .01)}

\item{add_quad}{Indicates whether to add a quadratic term to the linear ridge
regression (Default is TRUE)}

\item{internal_scaling}{Specifies the type of standardization that should be 
applied internally to x and y. This is common in ESN literature.
Options are "joint" (the mean and standard deviation from all values 
in the x/y matrix are computed after appropriate times are removed
and used to center and scale the values in the x/y matrix) and "none"
(no internal scaling is applied). Default is "joint".}

\item{seed}{Random seed}

\item{n_ensembles}{Number of ensembles}

\item{cores}{Number of cores for parallelization (generally only needed
for large or deep ESNs)}
}
\description{
Function for training an ensemble ESN model.
}
\examples{
# Create data
x = matrix(c(rnorm(12,10,1), rnorm(12,0,1)), ncol = 2, byrow = FALSE)
y = matrix(x[,1], ncol = 1)

# Assign column names to data
colnames(x) = c("X1", "X2")
colnames(y) = "X1"

# Create a vector of times
t = c("Jan", "Feb", "Mar", "Apr", "May", "Jun",
      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
t = paste0(t, "2021")

# Fit an ESN model to training data
esn <- 
  fit_Eesn(
    x = x, 
    y = y, 
    t = t, 
    tau = 2, 
    m = 1, 
    tau_emb = 1, 
    nh = 50, 
    seed = 102034985,
    n_ensembles = 10,
    cores = 1
  )
}
