% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ridge_reg.R
\name{fit_ridge_reg}
\alias{fit_ridge_reg}
\title{Estimate ridge regression parameters}
\usage{
fit_ridge_reg(h_obj, add_quad, reg_par, nh, data_obj, internal_scaling)
}
\arguments{
\item{h_obj}{Object output from `create_h`}

\item{add_quad}{Indicates whether to add a quadratic term to the linear ridge
regression}

\item{reg_par}{Output ridge regression parameter}

\item{nh}{Number of hidden units}

\item{data_obj}{Output from create_embedded_data}

\item{internal_scaling}{Specifies the type of standardization that should be 
applied internally to x and y. This is common in ESN literature.
Options are "joint" (the mean and standard deviation from all values 
in the x/y matrix are computed after appropriate times are removed
and used to center and scale the values in the x/y matrix) and "none"
(no internal scaling is applied). Default is "joint".}
}
\description{
This function returns the matrix of ridge regression parameters, V.
}
