% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_search.R
\name{hyperparameter_search}
\alias{hyperparameter_search}
\title{Fit Ensembled (bagged) ESN model}
\usage{
hyperparameter_search(
  x,
  y,
  t,
  x_test,
  t_test,
  phi,
  obs_train,
  obs_test,
  tau,
  m,
  tau_emb,
  nh,
  U_width,
  W_width,
  U_pi,
  W_pi,
  nu,
  reg_par,
  add_quad,
  internal_scaling = "joint",
  weights = NULL,
  seed,
  n_ensembles = 1,
  cores = 1
)
}
\arguments{
\item{x}{Matrix of training data inputs with T rows (number of times) and 
P columns (number of input variables). Rows must be ordered 
top to bottom from earliest to latest observation.}

\item{y}{Matrix of training data outputs with T rows (number of times) and 
Q columns (number of response variables). Times associated
with the rows must match those in x.}

\item{t}{Vector containing times/dates associated with x and y matrices.}

\item{x_test}{Matrix of testing data inputs with T' rows (number of times) 
and P columns (number of input variables). Rows must be ordered 
top to bottom from earliest to latest observation.}

\item{t_test}{Vector containing times/dates associated with x_test.}

\item{phi}{Matrix object phi output from compute_eofs when applied to Ztrain.}

\item{obs_train}{Isn't this the same as x?}

\item{obs_test}{Isn't this the same as x_test?}

\item{tau}{Forecast lead time}

\item{m}{Embedding vector length}

\item{tau_emb}{Embedding lag}

\item{nh}{Number of hidden units}

\item{U_width}{Uniform distribution sampling range for U ("width" parameter)
(Default is .1)}

\item{W_width}{Uniform distribution sampling range for W ("width" parameter)
(Default is .1)}

\item{U_pi}{Probabilities of non-zeros in U (sparseness parameter)
(Default is .1)}

\item{W_pi}{Probabilities of non-zeros in W (sparseness parameter)
(Default is .1)}

\item{nu}{Scaling parameter for W-weight matrix) (Default is .35)}

\item{reg_par}{Output ridge regression parameter (Default is .01)}

\item{add_quad}{Indicates whether to add a quadratic term to the linear 
ridge regression (Default is TRUE)}

\item{internal_scaling}{Specifies the type of standardization that should be 
applied internally to x and y. This is common in ESN literature.
Options are "joint" (the mean and standard deviation from all values 
in the x/y matrix are computed after appropriate times are removed
and used to center and scale the values in the x/y matrix) and "none"
(no internal scaling is applied). Default is "joint".}

\item{weights}{Vector of weights for computing weighted RMSEs (if NULL, 
RMSEs are computed without weighting)}

\item{seed}{Random seed}

\item{n_ensembles}{number of ensembles}

\item{cores}{number of cores for parallelization (generally only needed for
large or deep ESNs)}
}
\description{
Function for training an Ensembled (bagged) ESN model.
}
\examples{
# Create training data
x = matrix(c(rnorm(12,10,1), rnorm(12,0,1)), ncol = 2, byrow = FALSE)
y = matrix(c(rnorm(12,5,3), rnorm(12,5,3)), ncol = 2, byrow = FALSE)

# Create a vector of times
t = c("Jan", "Feb", "Mar", "Apr", "May", "Jun",
      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
t = paste0(t, "2021")

# Create test data
x_new = matrix(c(rnorm(5,10,1), rnorm(5,0,1)), ncol = 2, byrow = FALSE)
t_new = c("Jan", "Feb", "Mar", "Apr", "May")
t_new = paste0(t_new, "2022")

# Assign column names to data
colnames(x) = c("X1", "X2")
colnames(y) = c("Y1", "Y2")
colnames(x_new) = c("X1", "X2")

# Compute EOFs
x_eofs <-
  compute_eofs(
    Ztrain = x,
    Ztest = x_new,
    n_eofs = 2
   )
y_eofs <-
  compute_eofs(
    Ztrain = y,
    n_eofs = 2
   )
   
# Implement hyperparameter search
rmses <- 
  hyperparameter_search(
    x = x_eofs$train,
    y = y_eofs$train,
    t = t,
    x_test = x_eofs$test,
    t_test = t_new,
    phi = x_eofs$phi,
    obs_train = x,
    obs_test = x_new,
    tau = 1,
    m = 1, 
    tau_emb = 1, 
    nh = 50, 
    U_width = 0.1,
    W_width = 0.1, 
    U_pi = 0.1, 
    W_pi = 0.1, 
    nu = c(0.5,1), 
    reg_par = c(1,2), 
    add_quad = TRUE, 
    internal_scaling = "joint", 
    seed = 1020349858
  )
}
