% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merra2.R
\docType{data}
\name{merra2}
\alias{merra2}
\title{merra2 dataset}
\format{
A data.frame.
}
\usage{
merra2
}
\description{
This dataset contains data from the Modern-Era Retrospective Analysis for 
Research and Applications, Version 2 (MERRA-2). See 
doi:10.1175/JCLI-D-16-0758.1 for details on MERRA-2. Both stratospheric 
temperature (50 mb) and vertically integrated AOD contain monthly values
from 1980 to 1995. For the code used to prepare this dataset, see
https://github.com/sandialabs/listenr/inst/package-data-prep/prep-merra2.md.
}
\keyword{datasets}
