% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_esn.R
\name{predict_esn}
\alias{predict_esn}
\title{Obtain predictions/forecasts with ESN model}
\usage{
predict_esn(model, x_new = NULL, t_new = NULL, x_ood = NULL, phi = NULL)
}
\arguments{
\item{model}{Object output from fit_esn function}

\item{x_new}{Data input matrix occurring after the times in the training
data x}

\item{t_new}{Vector of times associated with x_new matrix}

\item{x_ood}{Out of distribution data inputs (must be the same size matrix
as training data x and occur at the same times as x)}

\item{phi}{Matrix object phi output from compute_eofs when applied to 
Ztrain. If specified, predictions will be returned on spatial scale.
Default is NULL.}
}
\value{
List. (See details for more information.)
}
\description{
This function computes predictions for a model trained using `fit_esn`. Both
predictions on the training data and out-of-sample forecasts on new data can
be computed. New data must sequentially follow training data (no gaps in
time).
}
\details{
Objects output from `predict_esn` when no new `x` values are input:
\itemize{
  \item \code{preds_ins}: Predictions computed using "in-sample" data used to train the model
  \item \code{preds_oos}: Forecasts computed using x data input to `fit_esn` but not used to train the model (i.e., "out-of-sample" data)
  \item \code{data_oos}: Data object created for the out-of-sample observations to obtain `preds_oos`
  \item \code{h_oos}: h matrix object created for the out-of-sample observations to obtain `preds_oos`
}

Objects output from `predict_esn` when new `x` values are input:
\itemize{
  \item \code{preds_new}: Forecasts computed using `x_new` data input to `predict_esn`
  \item \code{data_new}: Data object created for the new observations to obtain `preds_new`
  \item \code{h_new}: h matrix object created for the new observations to obtain `preds_new`
}
}
\examples{
# Create data
x = matrix(c(rnorm(12,10,1), rnorm(12,0,1)), ncol = 2, byrow = FALSE)
y = matrix(x[,1], ncol = 1)

# Assign column names to data
colnames(x) = c("X1", "X2")
colnames(y) = "X1"

# Create a vector of times
t = c("Jan", "Feb", "Mar", "Apr", "May", "Jun", 
      "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
t = paste0(t, "2021")

# Fit an ESN model to training data
esn <- 
  fit_esn(
    x = x, 
    y = y, 
    t = t, 
    tau = 2, 
    m = 1, 
    tau_emb = 1, 
    nh = 50, 
    seed = 1020349858
   )

# Get predictions on in-sample training data and forecasts on out-of-sample 
# training data
preds = predict_esn(model = esn)

# Get forecasts on new data
x_new = matrix(c(rnorm(5,10,1), rnorm(5,0,1)), ncol = 2, byrow = FALSE)
t_new = c("Jan", "Feb", "Mar", "Apr", "May")
t_new = paste0(t_new, "2022")
forecasts = predict_esn(model = esn, x_new = x_new, t_new = t_new)

# Get predictions on new (OOD) data
x_ood = matrix(c(rcauchy(12, 7, 1), rcauchy(12, 2, 1)), ncol = 2, byrow = FALSE)
new_pred = predict_esn(model = esn, x_ood = x_ood)
}
