% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulated dataset}
\format{
A data.frame.
}
\usage{
sim
}
\description{
A simulated dataset generate for `listenr` examples. For the code used to 
prepare this dataset, see
https://github.com/sandialabs/listenr/inst/package-data-prep/prep-sim-data.md.
}
\keyword{datasets}
