% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_synthetic_data.R
\name{simulate_synthetic_data}
\alias{simulate_synthetic_data}
\title{Function for simulating simple synthetic data from a linear model with
spatio-temporal covariates, spatio-temporal noise, and white noise}
\usage{
simulate_synthetic_data(
  nT,
  nx,
  ny,
  x1mean,
  x2mean,
  xsd,
  sigmaX,
  sigmaDelta,
  sigmaEpsilon,
  phiX,
  phiDelta,
  rhoX,
  rhoDelta,
  seed = 1234
)
}
\arguments{
\item{nT}{number of times points}

\item{nx}{number of x values in grid}

\item{ny}{number of y values in grid}

\item{x1mean}{location of mode for x1}

\item{x2mean}{location of mode for x2}

\item{xsd}{standard deviation for covariates}

\item{sigmaX}{sqrt of sill of covariates (sqrt(covariance) at distance=0)}

\item{sigmaDelta}{sqrt of sill of spatio-temporal random effect (sqrt(covariance) at distance=0)}

\item{sigmaEpsilon}{sd of white noise}

\item{phiX}{range of covariates (sqrt(covariance) at distance=0)}

\item{phiDelta}{range of spatio-temporal random effect (sqrt(covariance) at distance=0)}

\item{rhoX}{AR(1) coefficient for temporal correlation of covariates}

\item{rhoDelta}{AR(1) coefficient for temporal correlation of spatio-temporal random effect}

\item{seed}{Random seed}
}
\description{
Covariate X2 has real effect on response, covariate X1 has no effect on response
}
\examples{
# Simulate data
dat = simulate_synthetic_data(nT = 70, nx = 10, ny = 10, x1mean = 20, 
                              x2mean = 45, xsd = 6, sigmaX = 2, 
                              sigmaDelta = 2, sigmaEpsilon = 2, 
                              phiX = .2, phiDelta = .2, rhoX = .2, 
                              rhoDelta = .2)
}
